backend/requirements-api.txt
# Voice Guard — API 서버 전용 의존성
# openai-whisper(PyTorch, ~2GB)는 worker.py 전용 → 이 파일에서 제외

fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.18
pydantic==2.10.3
requests==2.32.3
httpx==0.28.1
pandas>=2.2.0
openpyxl>=3.1.0
boto3==1.35.76
redis>=5.0.0
firebase-admin>=6.5.0
anthropic>=0.40.0
google-generativeai>=0.8.0
openai>=1.0.0
